refactor: Split up lookup.go
